requirements.txt
aiogram
python-dotenv
flask
flask-cors
python-dotenv